+++ lib/common/constants/image_assets.dart
@@ -1,12 +1,19 @@
 abstract class ImageAssets {
   static const splash = 'assets/img/bg_splash.png';
-  static const document = 'assets/icons/what_get_icon.svg';
-  static const backArrow = 'assets/icons/back_arrow.svg';
-  static const search = 'assets/icons/search.svg';
-  static const filter = 'assets/icons/filter.svg';
-  static const phone = 'assets/icons/phone.svg';
-  static const message = 'assets/icons/message.svg';
-  static const email = 'assets/icons/email.svg';
-  static const checkmark = 'assets/icons/check_mark.svg';
 
+  static const iconsPath = 'assets/icons';
+  static const logo = '$iconsPath/logo.svg';
+  static const document = '$iconsPath/what_get_icon.svg';
+  static const backArrow = '$iconsPath/back_arrow.svg';
+  static const search = '$iconsPath/search.svg';
+  static const filter = '$iconsPath/filter.svg';
+  static const phone = '$iconsPath/phone.svg';
+  static const message = '$iconsPath/message.svg';
+  static const email = '$iconsPath/email.svg';
+  static const checkmark = '$iconsPath/check_mark.svg';
+  static const calendar = '$iconsPath/calendar.svg';
+  static const flyVocation = '$iconsPath/fly_vocation.svg';
+  static const lock = '$iconsPath/lock.svg';
+  static const payList = '$iconsPath/pay_list.svg';
+  static const support = '$iconsPath/support.svg';
 }
