+++ lib/feature/domain/entities/declaration_entity.dart
@@ -1,0 +1,60 @@
+class DeclarationEntity {
+  final String title;
+  final String svgPath;
+  final String date;
+  final String number;
+  final String status;
+
+  DeclarationEntity({
+    required this.title,
+    required this.svgPath,
+    required this.date,
+    required this.number,
+    required this.status,
+  });
+}
+
+final List<DeclarationEntity> mockDeclarations = [
+  DeclarationEntity(
+    title: 'Заявление на отпуск',
+    svgPath: 'assets/icons/fly_vocation.svg',
+    date: '17 августа 2022 15:34:56',
+    number: '#И213212111',
+    status: 'обработка',
+  ),
+  DeclarationEntity(
+    title: 'Заявление на командировку',
+    svgPath: 'assets/icons/calendar.svg',
+    date: '17 августа 2022 15:34:56',
+    number: '#И213212111',
+    status: 'одобрено',
+  ),
+  DeclarationEntity(
+    title: 'Заявление на пропуск',
+    svgPath: 'assets/icons/lock.svg',
+    date: '17 августа 2022 15:34:56',
+    number: '#И213212111',
+    status: 'обработка',
+  ),
+  DeclarationEntity(
+    title: 'Заявление на расчетный листок',
+    svgPath: 'assets/icons/pay_list.svg',
+    date: '17 августа 2022 15:34:56',
+    number: '#И213212111',
+    status: 'одобрено',
+  ),
+  DeclarationEntity(
+    title: 'Заявление на тех. поддержку/IT',
+    svgPath: 'assets/icons/support.svg',
+    date: '17 августа 2022 15:34:56',
+    number: '#И213212111',
+    status: 'отклонено',
+  ),
+  DeclarationEntity(
+    title: 'Заявление на тех. поддержку/IT',
+    svgPath: 'assets/icons/support.svg',
+    date: '17 августа 2022 15:34:56',
+    number: '#И213212111',
+    status: 'одобрено',
+  ),
+];
